Encountered weird sb error

+++ java/src/main/java/com/techelevator/dao/jdbcdao/JdbcInsulinDao.java
@@ -21,7 +21,7 @@ public class JdbcInsulinDao implements InsulinDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    // get single insulin object
+    // get single insulin object -done
     // get list of active insulin objects - done
     // create a new insulin data point - done
     // update insulin attributes - done
@@ -32,14 +32,15 @@ public class JdbcInsulinDao implements InsulinDao {
 
         List<BaseInsulin> insulinList = new ArrayList<>();
 
-        String sql = "SELECT i.insulin_id, i.base_level, i.avg_level, i.time_last_dose, i.insulin_type, i.insulin_strength, i.half_life, i.onset, i.peak, i.insulin_ration, i.duration FROM insulin i " +
-                "JOIN insulin_user_data_join ij ON ij.insulin_id = i.insulin_id " +
+        String sql = "SELECT i.insulin_id, base_level, avg_level, time_last_dose, insulin_brand_name, insulin_strength, insulin_ratio " +
+                "FROM insulin i " +
+                "JOIN insulin_user_data_join ij ON i.insulin_id = ij.insulin_id " +
                 "JOIN user_data ud ON ud.user_id = ij.user_id " +
                 "WHERE ud.user_id = ?;";
 
         SqlRowSet rowSet = jdbcTemplate.queryForRowSet(sql, userId);
         while (rowSet.next()) {
-            insulinList.add(mapToRowSet(rowSet));
+            insulinList.add(mapToRowSetBaseInsulin(rowSet));
         }
         return insulinList;
     }
@@ -47,10 +48,11 @@ public class JdbcInsulinDao implements InsulinDao {
     @Override
     public BaseInsulin createNewInsulin(int userId, BaseInsulin baseInsulin) throws SQLException {
 
-        String sql = "INSERT INTO insulin (base_level, avg_level, time_last_dose, insulin_type, insulin_strength, insulin_ration) " +
+        String sql = "INSERT INTO insulin (base_level, avg_level, time_last_dose, insulin_brand_name, insulin_strength, insulin_ratio)\n" +
                 "VALUES (?, ?, ?, ?, ?, ?) RETURNING insulin_id;";
 
-        Integer id = jdbcTemplate.queryForObject(sql, Integer.class, baseInsulin.getBaseLevel(), baseInsulin.getAverageLevel(), baseInsulin.getTimeSinceLastDose(), baseInsulin.getInsulinType(), baseInsulin.getInsulinStrength(), baseInsulin.getInsulinRation());
+        Integer id = jdbcTemplate.queryForObject(sql, Integer.class, baseInsulin.getBaseLevel(), baseInsulin.getAverageLevel(), baseInsulin.getTimeSinceLastDose(),
+                baseInsulin.getInsulinBrandName(), baseInsulin.getInsulinStrength(), baseInsulin.getInsulinRatio());
         baseInsulin.setInsulinId(id);
 
         if (createNewJoinInsulinEntry(userId, id)) {
@@ -64,11 +66,11 @@ public class JdbcInsulinDao implements InsulinDao {
     @Override
     public boolean updateInsulin(BaseInsulin baseInsulin) throws SQLException {
 
-        String sql = "UPDATE insulin SET base_level = ?, avg_level = ?, " +
-                "time_last_dose = ?, insulin_type = ?, insulin_strength = ?, insulin_ration = ? WHERE insulin_id = ?;";
+        String sql = "UPDATE insulin SET base_level = ?, avg_level = ? time_last_dose = ?, insulin_brand_name = ?, " +
+                "insulin_strength = ?, insulin_ratio = ? WHERE insulin_id = ?;";
 
         int success = jdbcTemplate.update(sql, baseInsulin.getBaseLevel(), baseInsulin.getAverageLevel(), baseInsulin.getTimeSinceLastDose(),
-                baseInsulin.getInsulinType(), baseInsulin.getInsulinStrength(), baseInsulin.getInsulinRation(), baseInsulin.getInsulinId());
+                baseInsulin.getInsulinStrength(), baseInsulin.getInsulinRatio(), baseInsulin.getInsulinId());
 
         if (success == 1) {
             return true;
@@ -94,17 +96,16 @@ public class JdbcInsulinDao implements InsulinDao {
     @Override
     public BaseInsulin getSingleInsulin(int userId, int insulinId, BaseInsulin insulin) throws SQLException {
 
-        String sql = "SELECT insulin.insulin_id, base_level, avg_level, time_last_dose, insulin_type, insulin_strength, half_life, onset, peak, insulin_ration, duration " +
-                "FROM insulin " +
-                "JOIN insulin_data_join ij ON ij.insulin_id = insulin.insulin_id " +
+        String sql = "SELECT i.insulin_id, base_level, avg_level, time_last_dose, insulin_brand_name, insulin_strength, insulin_ratio " +
+                "FROM insulin i " +
+                "JOIN insulin_user_data_join ij ON i.insulin_id = ij.insulin_id " +
                 "JOIN user_data ud ON ud.user_id = ij.user_id " +
-                "WHERE ud.user_id = ? AND insulin.insulin_id = ?";
+                "WHERE ud.user_id = ? AND i.insulin_id = ?;";
 
         SqlRowSet rowSet = jdbcTemplate.queryForRowSet(sql, userId, insulin);
 
         if (rowSet.next()) {
-            BaseInsulin baseInsulin = mapToRowSetBaseInsulin(rowSet);
-            return baseInsulin;
+            return mapToRowSetBaseInsulin(rowSet);
         } else {
             throw new SQLException("Cannot find single Insulin object");
         }
@@ -124,37 +125,40 @@ public class JdbcInsulinDao implements InsulinDao {
 
     }
 
-    private void deleteFromJoinTable(int insulinId) {
-        String sql = "DELETE FROM insulin_user_data_join WHERE insulin_id = ?;";
-        jdbcTemplate.update(sql, insulinId);
+    private void deleteFromJoinTable(int insulinId) throws SQLException {
+        String sql = "DELETE FROM insulin_user_data_join WHERE insulin_id = ? AND user_id = ?;";
+        int rowsAffected = jdbcTemplate.update(sql, insulinId);
+        if (rowsAffected != 1) {
+            throw new SQLException("Delete failed.");
+        }
     }
 
-    private InsulinInformation mapToRowSet(SqlRowSet row) {
+    private InsulinInformation mapToRowSetInsulinInformation(SqlRowSet rowSet) {
         InsulinInformation insulin = new InsulinInformation();
-        insulin.setInsulinId(row.getInt("insulin_id"));
-        insulin.setBaseLevel(row.getDouble("base_level"));
-        insulin.setAverageLevel(row.getDouble("avg_level"));
-        insulin.setTimeSinceLastDose(row.getTimestamp("time_last_dose"));
-        insulin.setInsulinType(row.getString("insulin_type"));
-        insulin.setInsulinStrength(row.getString("insulin_strength"));
-        insulin.setHalfLife(row.getInt("half_life"));
-        insulin.setOnset(row.getInt("onset"));
-        insulin.setPeak(row.getInt("peak"));
-        insulin.setInsulinRation(row.getDouble("insulin_ration"));
-        insulin.setDuration(row.getInt("duration"));
+        fillSharedRows(rowSet, insulin);
+        insulin.setInsulinType(rowSet.getString("insulin_type"));
+        insulin.setHalfLife(rowSet.getInt("half_life"));
+        insulin.setOnsetLow(rowSet.getInt("onset_low"));
+        insulin.setOnsetLow(rowSet.getInt("onset_high"));
+        insulin.setPeak(rowSet.getInt("peak"));
+        insulin.setDuration(rowSet.getInt("duration"));
         return insulin;
     }
 
     private BaseInsulin mapToRowSetBaseInsulin(SqlRowSet rowSet) {
         BaseInsulin insulin = new BaseInsulin();
+        fillSharedRows(rowSet, insulin);
+        return insulin;
+    }
+
+    private void fillSharedRows(SqlRowSet rowSet, BaseInsulin insulin) {
         insulin.setInsulinId(rowSet.getInt("insulin_id"));
         insulin.setBaseLevel(rowSet.getDouble("base_level"));
         insulin.setAverageLevel(rowSet.getDouble("avg_level"));
         insulin.setTimeSinceLastDose(rowSet.getTimestamp("time_last_dose"));
-        insulin.setInsulinType(rowSet.getString("insulin_type"));
+        insulin.setInsulinBrandName(rowSet.getString("insulin_brand_name"));
         insulin.setInsulinStrength(rowSet.getString("insulin_strength"));
-        insulin.setInsulinRation(rowSet.getDouble("insulin_ration"));
-        return insulin;
+        insulin.setInsulinRatio(rowSet.getDouble("insulin_ratio"));
     }
 
 }
